Refactor: Complete code audit and architectural improvements.
This commit addresses the findings of a comprehensive code audit of the Travel Packing Agent and related components. The focus was on improving code quality, removing redundancy, enhancing modularity, and fixing bugs.

Key changes include:

1.  **Redundancy Elimination:**
    *   Removed the duplicate `_prepare_travel_context` method from `core/travel_packing_agent.py`.
    *   Centralized the `_categorize_items_for_travel` function into a new `core/utils.py` module to be used by both the agent and the orchestrator.

2.  **Architectural Refactoring:**
    *   Created a new `PromptBuilder` class in `core/prompt_builder.py` to encapsulate all prompt and text generation logic, significantly reducing the size and complexity of the `TravelPackingAgent`.
    *   Moved all prompt building, destination tip generation, and example outfit generation logic to the `PromptBuilder`, cleaning up the agent's responsibilities.

3.  **Code Cleanup:**
    *   Removed numerous unused imports from `core/travel_packing_agent.py` and `core/ai_packing_optimizer.py`.
    *   Removed the dead code method `_format_items_concise` from the agent.

4.  **Bug Fixes:**
    *   Corrected a typo in `__main__.py` that caused the environment variable check to fail silently.
    *   Fixed a bug in `data/notion_utils.py` where a function was using the wrong status string ("Done" instead of "washed"), causing a test to fail.

All existing tests pass after this extensive refactoring, ensuring that the core functionality remains intact.
