+++ data/notion_utils.py
@@ -676,8 +676,8 @@ def remove_from_dirty_clothes_and_mark_washed(dirty_item_page_id: str):
         logging.info(f"Removed page {dirty_item_page_id} from Dirty Clothes database")
         
         # Update the original clothing item's washed status to "Done"
-        update_clothing_washed_status(clothing_item_id, "Done")
-        logging.info(f"Marked clothing item {clothing_item_id} as Done")
+        update_clothing_washed_status(clothing_item_id, "washed")
+        logging.info(f"Marked clothing item {clothing_item_id} as washed")
         
         return True
